[ADD] 23/10/23 9:30 모꾸 게시물, 댓글 무한스크롤 적용, 짜잘한 css수정
1. 모꾸 게시물, 댓글리스트에 무한 스크롤 적용함
2. 지도 마커 커스텀오버레이창 css 일부 적용
3. 게시물 css 수정
4. 모꾸 서치 구현 중

+++ src/main/java/com/jiqoo/moqoo/controller/MoqooComtController.java
@@ -30,18 +30,12 @@ public class MoqooComtController {
 	// 댓글 작성
 	@ResponseBody
 	@PostMapping("/comt/insert")
-	public String insertComt(@RequestParam("comtContent") String comtContent
-			, @RequestParam("refMoqooNo") int refMoqooNo
-			, HttpSession session) {
+	public String insertComt(@ModelAttribute Comment comt, HttpSession session) {
 		String comtWriter = (String)session.getAttribute("userId");
-		Comment comt = new Comment();
-		comt.setRefPostNo(refMoqooNo);
-		comt.setComtContent(comtContent);
-		int result = 0;
-		if(comtWriter != null && !comtWriter.equals("")) {
-			comt.setComtWriter(comtWriter);
-			result = moqooComtService.insertComt(comt);
-		}
+//		if(comtWriter != null && !comtWriter.equals("")) {
+		comt.setComtWriter(comtWriter);
+//		}
+		int result = moqooComtService.insertComt(comt);
 		if(result > 0) {
 			return "success";
 		}
@@ -55,12 +49,21 @@ public class MoqooComtController {
 	@ResponseBody
 	@GetMapping("/moqoo/delComt")
 	public String deleteComt(@ModelAttribute Comment comt) {
-		int result = moqooComtService.deleteComt(comt);
-		if(result > 0) {
-			return "success";
-		}
-		else {
-			return "fail";
+		int countChildComt = moqooComtService.countChildComment(comt);
+		if(countChildComt > 0) {
+			int updateDelComment = moqooComtService.updateDelComment(comt);
+			if(updateDelComment > 0) {
+				return "success";
+			}else {
+				return "fail";
+			}
+		}else {
+		   int result = moqooComtService.deleteComt(comt);
+		   if(result > 0) {
+			   return "success";
+		   }else {
+			   return "fail";
+		   }
 		}
 	}
